refactor: move frame_rate under the runtime module
`FrameRate`/`FrameRateError` are only consumed by the runtime (they are
the scheduling input to `Runtime::new` and `FrameScheduler`), yet lived
at the crate root next to unrelated top-level modules. That raised
recurring reader questions — why is it separate from `frame_scheduler`,
is the top-level placement intentional, should `frame_scheduler` move up?
Co-locate it with the scheduling types under `runtime/`.

- `src/frame_rate.rs` → `src/runtime/frame_rate.rs`, declared
  `pub(crate) mod frame_rate` in `runtime`.
- `lib.rs` and `prelude` re-export from `crate::runtime::frame_rate`, so
  the public API is unchanged: `tears::FrameRate` / `tears::FrameRateError`
  and their `prelude` forms remain the single canonical paths. `pub(crate)`
  keeps the module itself from becoming a second public path.
- `frame_scheduler` now imports it via `super::frame_rate`.

No API or behavior change.
